types: make security package strict-clean and remove it from the mypy ledger
Narrow the watermark trigger activation to `float`. Drop the `security.*` glob
from the mypy override ledger; the whole tree stays strict-clean.

Authored by Anulum Fortis & Arcane Sapience (protoscience@anulum.li)

+++ src/sc_neurocore/security/watermark.py
@@ -73,5 +73,5 @@ class WatermarkInjector:
         # SC Dot Product Ideal: Sum(x * w) / Length
         # Here we just check alignment
 
-        activation = np.mean(trigger_pattern * w)
+        activation = float(np.mean(trigger_pattern * w))
         return activation
